requirements.txt
ultralytics>=8.3.57
flask>=3.1.0
pillow>=11.0.0
Flask-Cors==5.0.0
